pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2md"
version = "0.1.0"
description = "Convert PDFs to Markdown with LaTeX math support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
keywords = ["pdf", "markdown", "latex", "ocr", "converter", "obsidian"]
authors = [
    {name = "DEVLlN"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "marker-pdf>=1.0.0",
    "pymupdf>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
pdf2md = "pdf2md:main"

[project.urls]
Homepage = "https://github.com/DEVLlN/pdf2md"
Repository = "https://github.com/DEVLlN/pdf2md"
Issues = "https://github.com/DEVLlN/pdf2md/issues"

[tool.setuptools]
py-modules = ["pdf2md"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
